ffmpeg: move output open/reopen to encoder module
