+++ ffmpeg/transcoder.c
@@ -6,9 +6,6 @@
 #include <libavcodec/avcodec.h>
 #include <libavformat/avformat.h>
 
-#include <libavfilter/buffersrc.h>
-#include <libavfilter/buffersink.h>
-
 // Not great to appropriate internal API like this...
 const int lpms_ERR_INPUT_PIXFMT = FFERRTAG('I','N','P','X');
 const int lpms_ERR_INPUT_CODEC = FFERRTAG('I','N','P','C');
@@ -89,258 +86,11 @@ void lpms_init()
 // Transcoder
 //
 
-static void close_output(struct output_ctx *octx)
-{
-  if (octx->oc) {
-    if (!(octx->oc->oformat->flags & AVFMT_NOFILE) && octx->oc->pb) {
-      avio_closep(&octx->oc->pb);
-    }
-    avformat_free_context(octx->oc);
-    octx->oc = NULL;
-  }
-  if (octx->vc && AV_HWDEVICE_TYPE_NONE == octx->hw_type) avcodec_free_context(&octx->vc);
-  if (octx->ac) avcodec_free_context(&octx->ac);
-  octx->af.flushed = octx->vf.flushed = 0;
-  octx->af.flushing = octx->vf.flushing = 0;
-  octx->vf.pts_diff = INT64_MIN;
-}
-
-static void free_output(struct output_ctx *octx) {
-  close_output(octx);
-  if (octx->vc) avcodec_free_context(&octx->vc);
-  free_filter(&octx->vf);
-  free_filter(&octx->af);
-}
-
 static int is_mpegts(AVFormatContext *ic) {
   return !strcmp("mpegts", ic->iformat->name);
 }
 
-static int add_video_stream(struct output_ctx *octx, struct input_ctx *ictx)
-{
-#define vs_err(msg) { \
-  if (!ret) ret = -1; \
-  fprintf(stderr, "Error adding video stream: " msg); \
-  goto add_video_err; \
-}
-
-  // video stream to muxer
-  int ret = 0;
-  AVStream *st = avformat_new_stream(octx->oc, NULL);
-  if (!st) vs_err("Unable to alloc video stream\n");
-  octx->vi = st->index;
-  st->avg_frame_rate = octx->fps;
-  if (is_copy(octx->video->name)) {
-    AVStream *ist = ictx->ic->streams[ictx->vi];
-    if (ictx->vi < 0 || !ist) vs_err("Input video stream does not exist\n");
-    st->time_base = ist->time_base;
-    ret = avcodec_parameters_copy(st->codecpar, ist->codecpar);
-    if (ret < 0) vs_err("Error copying video params from input stream\n");
-    // Sometimes the codec tag is wonky for some reason, so correct it
-    ret = av_codec_get_tag2(octx->oc->oformat->codec_tag, st->codecpar->codec_id, &st->codecpar->codec_tag);
-    avformat_transfer_internal_stream_timing_info(octx->oc->oformat, st, ist, AVFMT_TBCF_DEMUXER);
-  } else if (octx->vc) {
-    st->time_base = octx->vc->time_base;
-    ret = avcodec_parameters_from_context(st->codecpar, octx->vc);
-    if (octx->gop_time) {
-      // Rescale the gop time to the expected timebase after filtering.
-      // The FPS filter outputs pts incrementing by 1 at a rate of 1/framerate
-      // while non-fps will retain the input timebase.
-      AVRational gop_tb = {1, 1000};
-      AVRational dest_tb;
-      if (octx->fps.den) dest_tb = av_inv_q(octx->fps);
-      else dest_tb = ictx->ic->streams[ictx->vi]->time_base;
-      octx->gop_pts_len = av_rescale_q(octx->gop_time, gop_tb, dest_tb);
-      octx->next_kf_pts = 0; // force for first frame
-    }
-    if (ret < 0) vs_err("Error setting video params from encoder\n");
-  } else vs_err("No video encoder, not a copy; what is this?\n");
-  return 0;
-
-add_video_err:
-  // XXX free anything here?
-  return ret;
-#undef vs_err
-}
-
-static int add_audio_stream(struct input_ctx *ictx, struct output_ctx *octx)
-{
-#define as_err(msg) { \
-  if (!ret) ret = -1; \
-  fprintf(stderr, "Error adding audio stream: " msg); \
-  goto add_audio_err; \
-}
-
-  if (ictx->ai < 0 || octx->da) {
-    // Don't need to add an audio stream if no input audio exists,
-    // or we're dropping the output audio stream
-    return 0;
-  }
-
-  // audio stream to muxer
-  int ret = 0;
-  AVStream *st = avformat_new_stream(octx->oc, NULL);
-  if (!st) as_err("Unable to alloc audio stream\n");
-  if (is_copy(octx->audio->name)) {
-    AVStream *ist = ictx->ic->streams[ictx->ai];
-    if (ictx->ai < 0 || !ist) as_err("Input audio stream does not exist\n");
-    st->time_base = ist->time_base;
-    ret = avcodec_parameters_copy(st->codecpar, ist->codecpar);
-    if (ret < 0) as_err("Error copying audio params from input stream\n");
-    // Sometimes the codec tag is wonky for some reason, so correct it
-    ret = av_codec_get_tag2(octx->oc->oformat->codec_tag, st->codecpar->codec_id, &st->codecpar->codec_tag);
-    avformat_transfer_internal_stream_timing_info(octx->oc->oformat, st, ist, AVFMT_TBCF_DEMUXER);
-  } else if (octx->ac) {
-    st->time_base = octx->ac->time_base;
-    ret = avcodec_parameters_from_context(st->codecpar, octx->ac);
-    if (ret < 0) as_err("Error setting audio params from encoder\n");
-  } else if (is_drop(octx->audio->name)) {
-    // Supposed to exit this function early if there's a drop
-    as_err("Shouldn't ever happen here\n");
-  } else {
-    as_err("No audio encoder; not a copy; what is this?\n");
-  }
-  octx->ai = st->index;
-
-  // signal whether to drop preroll audio
-  if (st->codecpar->initial_padding) octx->drop_ts = AV_NOPTS_VALUE;
-  return 0;
-
-add_audio_err:
-  // XXX free anything here?
-  return ret;
-#undef as_err
-}
-
-static int open_audio_output(struct input_ctx *ictx, struct output_ctx *octx,
-  AVOutputFormat *fmt)
-{
-#define ao_err(msg) { \
-  if (!ret) ret = -1; \
-  fprintf(stderr, msg"\n"); \
-  goto audio_output_err; \
-}
-
-  int ret = 0;
-  AVCodec *codec = NULL;
-  AVCodecContext *ac = NULL;
-
-  // add audio encoder if a decoder exists and this output requires one
-  if (ictx->ac && needs_decoder(octx->audio->name)) {
-
-    // initialize audio filters
-    ret = init_audio_filters(ictx, octx);
-    if (ret < 0) ao_err("Unable to open audio filter")
-
-    // open encoder
-    codec = avcodec_find_encoder_by_name(octx->audio->name);
-    if (!codec) ao_err("Unable to find audio encoder");
-    // open audio encoder
-    ac = avcodec_alloc_context3(codec);
-    if (!ac) ao_err("Unable to alloc audio encoder");
-    octx->ac = ac;
-    ac->sample_fmt = av_buffersink_get_format(octx->af.sink_ctx);
-    ac->channel_layout = av_buffersink_get_channel_layout(octx->af.sink_ctx);
-    ac->channels = av_buffersink_get_channels(octx->af.sink_ctx);
-    ac->sample_rate = av_buffersink_get_sample_rate(octx->af.sink_ctx);
-    ac->time_base = av_buffersink_get_time_base(octx->af.sink_ctx);
-    if (fmt->flags & AVFMT_GLOBALHEADER) ac->flags |= AV_CODEC_FLAG_GLOBAL_HEADER;
-    ret = avcodec_open2(ac, codec, &octx->audio->opts);
-    if (ret < 0) ao_err("Error opening audio encoder");
-    av_buffersink_set_frame_size(octx->af.sink_ctx, ac->frame_size);
-  }
-
-  ret = add_audio_stream(ictx, octx);
-  if (ret < 0) ao_err("Error adding audio stream")
-
-audio_output_err:
-  // TODO clean up anything here?
-  return ret;
-
-#undef ao_err
-}
-
-
-static int open_output(struct output_ctx *octx, struct input_ctx *ictx)
-{
-#define em_err(msg) { \
-  if (!ret) ret = -1; \
-  fprintf(stderr, msg); \
-  goto open_output_err; \
-}
-  int ret = 0, inp_has_stream;
-
-  AVOutputFormat *fmt = NULL;
-  AVFormatContext *oc = NULL;
-  AVCodecContext *vc  = NULL;
-  AVCodec *codec      = NULL;
-
-  // open muxer
-  fmt = av_guess_format(octx->muxer->name, octx->fname, NULL);
-  if (!fmt) em_err("Unable to guess output format\n");
-  ret = avformat_alloc_output_context2(&oc, fmt, NULL, octx->fname);
-  if (ret < 0) em_err("Unable to alloc output context\n");
-  octx->oc = oc;
-
-  // add video encoder if a decoder exists and this output requires one
-  if (ictx->vc && needs_decoder(octx->video->name)) {
-    ret = init_video_filters(ictx, octx);
-    if (ret < 0) em_err("Unable to open video filter");
-
-    codec = avcodec_find_encoder_by_name(octx->video->name);
-    if (!codec) em_err("Unable to find encoder");
-
-    // open video encoder
-    // XXX use avoptions rather than manual enumeration
-    vc = avcodec_alloc_context3(codec);
-    if (!vc) em_err("Unable to alloc video encoder\n");
-    octx->vc = vc;
-    vc->width = av_buffersink_get_w(octx->vf.sink_ctx);
-    vc->height = av_buffersink_get_h(octx->vf.sink_ctx);
-    if (octx->fps.den) vc->framerate = av_buffersink_get_frame_rate(octx->vf.sink_ctx);
-    else vc->framerate = ictx->vc->framerate;
-    if (octx->fps.den) vc->time_base = av_buffersink_get_time_base(octx->vf.sink_ctx);
-    else if (ictx->vc->time_base.num && ictx->vc->time_base.den) vc->time_base = ictx->vc->time_base;
-    else vc->time_base = ictx->ic->streams[ictx->vi]->time_base;
-    if (octx->bitrate) vc->rc_min_rate = vc->rc_max_rate = vc->rc_buffer_size = octx->bitrate;
-    if (av_buffersink_get_hw_frames_ctx(octx->vf.sink_ctx)) {
-      vc->hw_frames_ctx =
-        av_buffer_ref(av_buffersink_get_hw_frames_ctx(octx->vf.sink_ctx));
-      if (!vc->hw_frames_ctx) em_err("Unable to alloc hardware context\n");
-    }
-    vc->pix_fmt = av_buffersink_get_format(octx->vf.sink_ctx); // XXX select based on encoder + input support
-    if (fmt->flags & AVFMT_GLOBALHEADER) vc->flags |= AV_CODEC_FLAG_GLOBAL_HEADER;
-    ret = avcodec_open2(vc, codec, &octx->video->opts);
-    if (ret < 0) em_err("Error opening video encoder\n");
-    octx->hw_type = ictx->hw_type;
-  }
-
-  // add video stream if input contains video
-  inp_has_stream = ictx->vi >= 0;
-  if (inp_has_stream && !octx->dv) {
-    ret = add_video_stream(octx, ictx);
-    if (ret < 0) em_err("Error adding video stream\n");
-  }
-
-  ret = open_audio_output(ictx, octx, fmt);
-  if (ret < 0) em_err("Error opening audio output\n");
-
-  if (!(fmt->flags & AVFMT_NOFILE)) {
-    ret = avio_open(&octx->oc->pb, octx->fname, AVIO_FLAG_WRITE);
-    if (ret < 0) em_err("Error opening output file\n");
-  }
-
-  ret = avformat_write_header(oc, &octx->muxer->opts);
-  if (ret < 0) em_err("Error writing header\n");
-
-  return 0;
-
-open_output_err:
-  free_output(octx);
-  return ret;
-}
-
-int flush_outputs(struct input_ctx *ictx, struct output_ctx *octx)
+static int flush_outputs(struct input_ctx *ictx, struct output_ctx *octx)
 {
   // only issue w this flushing method is it's not necessarily sequential
   // wrt all the outputs; might want to iterate on each output per frame?
@@ -430,29 +180,9 @@ int transcode(struct transcode_thread *h,
         continue;
       }
 
-      // non-first segment
-      // re-open muxer for HW encoding
-      AVOutputFormat *fmt = av_guess_format(octx->muxer->name, octx->fname, NULL);
-      if (!fmt) main_err("Unable to guess format for reopen\n");
-      ret = avformat_alloc_output_context2(&octx->oc, fmt, NULL, octx->fname);
-      if (ret < 0) main_err("Unable to alloc reopened out context\n");
-
-      // re-attach video encoder
-      if (octx->vc) {
-        ret = add_video_stream(octx, ictx);
-        if (ret < 0) main_err("Unable to re-add video stream\n");
-      } else fprintf(stderr, "no video stream\n");
-
-      // re-attach audio encoder
-      ret = open_audio_output(ictx, octx, fmt);
-      if (ret < 0) main_err("Unable to re-add audio stream\n");
-
-      if (!(fmt->flags & AVFMT_NOFILE)) {
-        ret = avio_open(&octx->oc->pb, octx->fname, AVIO_FLAG_WRITE);
-        if (ret < 0) main_err("Error re-opening output file\n");
-      }
-      ret = avformat_write_header(octx->oc, &octx->muxer->opts);
-      if (ret < 0) main_err("Error re-writing header\n");
+      // non-first segment of a HW session
+      ret = reopen_output(octx, ictx);
+      if (ret < 0) main_err("transcoder: Unable to re-open output for HW session");
   }
 
   av_init_packet(&ipkt);
